# Optimize Bytes::copy_from_slice